refactor(orchestration): extract WorkerSessionManager and its three pending-* maps
Moves the fourteen worker-session methods out of the facade verbatim, along
with the three in-memory maps they own: pendingWorkerSessions,
pendingLogicalTransportSessions, pendingParallelStarts. 4347 -> 3970 lines.

This is the only task in the split that moves instance state, so it is the
only one where a wiring mistake is invisible to the persisted-state oracle.
The maps close the TOCTOU window between the parallel-capacity gate and the
inner persist mutate; two manager instances would each keep their own
counter, both pass the gate for the last slot, and over-dispatch at capacity.

The plan's method assignment was incomplete. Six lines outside the fourteen
read or write the maps, in three methods that belong to other services
(registerExternalCoordinator, requestDelegateForHuman, requestDelegateFromRpc).
Rather than let those reach for fields that no longer exist, the manager
exposes exactly four accessors: hasPendingWorkerSession,
hasPendingLogicalTransportSession, claimParallelStart, releaseParallelStart.

All four are synchronous and never call kernel.mutate. Every call site but
releaseParallelStart already runs inside a mutate critical section, and mutate
is non-reentrant -- an async accessor routed through it would throw on the
first call to registerExternalCoordinator.

WorkerSessionDeps also picks up closeWorkerSession, which the plan's Pick
omitted; reconcileParallelSlots calls it. Forced by the type system, and
behaviour-neutral.

Verified by mutation: constructing a second manager and handing it
reconcileParallelSlots turns exactly two tests red -- the concurrency suite's
over-dispatch assertion and the source-regex singleton test. The 185-test
frozen oracle and all 24 golden fixtures stay green, which is precisely why
both of those tests exist.

All fourteen bodies confirmed byte-identical to their originals by script,
modulo the dropped `private` modifier.
